Animated Circular Preloader
Here, we have contain the code of amazing Animated Circular Prelaoder to make your website amazing. Basically this repository contains one HTML and one CSS file, It is too easy to create. we hope you like this code. If you want to see a full video tutorial of this then follow the link:- https://youtu.be/m0A-w62hn9c